MP2-1203 make all cofinancing numbers as decimals

See merge request interact/ems/ems!884

+++ jems-api/src/main/kotlin/io/cloudflight/jems/api/project/dto/partner/cofinancing/ProjectPartnerCoFinancingInputDTO.kt
@@ -1,9 +1,11 @@
 package io.cloudflight.jems.api.project.dto.partner.cofinancing
 
+import java.math.BigDecimal
+
 data class ProjectPartnerCoFinancingInputDTO(
 
     val fundType: ProjectPartnerCoFinancingFundType,
     val fundId: Long? = null,
-    val percentage: Int? = null,
+    val percentage: BigDecimal? = null,
 
-)
+    )
